+++ docker-images/kafka/exporter-scripts/kafka_exporter_run.sh
@@ -31,9 +31,9 @@ kafkaserver="--kafka.server="$KAFKA_EXPORTER_KAFKA_SERVER
 
 listenaddress="--web.listen-address=:9404"
 
-tls="--tls.enabled --tls.ca-file=/etc/kafka-exporter/cluster-ca-certs/ca.crt --tls.cert-file=/etc/kafka-exporter/kafka-exporter-certs/kafka-exporter.crt  --tls.key-file=/etc/kafka-exporter/kafka-exporter-certs/kafka-exporter.key"
+allgroups="--legacy.partitions"
 
-sasl="--no-sasl.handshake"
+tls="--tls.enabled --tls.ca-file=/etc/kafka-exporter/cluster-ca-certs/ca.crt --tls.cert-file=/etc/kafka-exporter/kafka-exporter-certs/kafka-exporter.crt  --tls.key-file=/etc/kafka-exporter/kafka-exporter-certs/kafka-exporter.key"
 
 # starting Kafka Exporter with final configuration
 cat <<EOT > /tmp/run.sh
@@ -44,8 +44,8 @@ $tls \
 $kafkaserver \
 $saramaenable \
 $listenaddress \
+$allgroups \
 $loglevel \
-$sasl \
 $version
 EOT